created user entity, user principle

+++ user-service/src/main/java/com/mymart/userservice/dao/RolesDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mymart.userservice.entity.Roles;
 
-public interface RolesDao extends JpaRepository<Roles, Integer>{
+public interface RolesDao extends JpaRepository<Roles, Long>{
 
 }
